+++ Data/Entitites/BaseEntity.cs
@@ -5,13 +5,13 @@ namespace Data.Entitites
     public abstract class BaseEntity
     {
         [Key]
-        public int Id { get; set; }
+        public long Id { get; set; }
 
-        public int CreatedBy { get; set; }
+        public long CreatedBy { get; set; }
 
         public DateTime? CreatedOn { get; set; } = DateTime.UtcNow;
 
-        public int UpdatedBy { get; set; }
+        public long UpdatedBy { get; set; }
         public DateTime? UpdatedOn { get; set; }
     }
 }
