+++ Exceptions/NotFound/NotFoundException.cs
@@ -9,7 +9,7 @@ namespace Exceptions.NotFound
 {
     public class NotFoundException : Exception
     {
-        public NotFoundException(int id, string className) : base($"The {className} with id {id} doesn't exist in the database.")
+        public NotFoundException(long id, string className) : base($"The {className} with id {id} doesn't exist in the database.")
         {
         }
     }
