+++ Repository/Contracts/IOrderRepository.cs
@@ -4,11 +4,11 @@ namespace Repository.Contracts
 {
     public interface IOrderRepository : IRepository
     {
-        Task<List<OrderDetailProduct>> GetOrderDetailProducts(int id);
+        Task<List<OrderDetailProduct>> GetOrderDetailProducts(long id);
         Task ComputeTotalPriceAsync(List<OrderDetailProduct> products);
         void ComputeTotalPriceForOrder(Order order);
         void CreateRangeOrder(Order order);
         void CreateUserOrder(Order order);
-        Task<Order> CompleteOrder(int id);
+        Task<Order> CompleteOrder(long id);
     }
 }
